+++ app/controllers/idv/scan_id_acuant_controller.rb
@@ -1,4 +1,3 @@
-# rubocop:disable Metrics/ClassLength
 module Idv
   class ScanIdAcuantController < ScanIdBaseController
     before_action :ensure_fully_authenticated_user_or_token_user_id
@@ -6,49 +5,34 @@ module Idv
     before_action :return_if_liveness_disabled, only: [:liveness]
     before_action :return_if_throttled_else_increment, only: [:liveness]
 
-    ACUANT_PASS = 1
-    SUBSCRIPTION_DATA = [{
-      'DocumentProcessMode': 2,
-      'Id': Figaro.env.acuant_assure_id_subscription_id,
-      'IsActive': true,
-      'IsDevelopment': Rails.env.development?,
-      'IsTrial': false,
-      'Name': '',
-      'StorePII': false,
-    }].freeze
-    CLASSIFICATION_DATA = {
-      'Type': {
-      },
-    }.freeze
     GOOD_DOCUMENT = { 'Result': 1, 'Fields': [{}] }.freeze
     USER_SESSION_FLOW_ID = 'idv/doc_auth_v2'.freeze
 
     def subscriptions
-      render_json SUBSCRIPTION_DATA
+      render_json ::Acuant::Subscriptions.new.call
     end
 
     def instance
       session[:scan_id] = {}
-      render_json(wrap_network_errors { assure_id.create_document })
+      render_json ::Acuant::Instance.new.call
     end
 
     def image
-      assure_id.instance_id = params[:instance_id]
-      result = assure_id.post_image(request.body.read, params[:side].to_i)
-      render_json result
+      render_json ::Acuant::Image.new(params[:instance_id]).
+                    call(request.body.read, params[:side])
     end
 
     def classification
-      # use service when we accept multiple document types and dynamically decide # of sides
-      render_json CLASSIFICATION_DATA
+      render_json ::Acuant::Classification.new.call
     end
 
     def document
-      instance_id = params[:instance_id]
-      assure_id.instance_id = instance_id
-      data = wrap_network_errors { assure_id.document }
-      return unless data
-      render_json process_data_and_store_pii_in_session_if_document_passes(data, instance_id)
+      data, instance_id, pii = ::Acuant::Document.new(params[:instance_id]).call(current_user)
+      if pii
+        scan_id_session[:instance_id] = instance_id
+        scan_id_session[:pii] = pii
+      end
+      render_json data
     end
 
     def field_image
@@ -56,14 +40,10 @@ module Idv
     end
 
     def liveness
-      live_face_image = process_selfie
-      return unless live_face_image
-      scan_id_session[:liveness_pass] = true
-
-      face_image_from_document = fetch_face_from_document
-      return unless face_image_from_document
-
-      check_face_match(face_image_from_document, live_face_image)
+      is_live, is_face_match = ::Acuant::Liveness.
+                               new(scan_id_session[:instance_id]).call(request.body.read)
+      scan_id_session[:liveness_pass] = is_live
+      scan_id_session[:facematch_pass] = is_face_match
       render_json({})
     end
 
@@ -72,71 +52,6 @@ module Idv
     end
 
     private
-
-    def check_face_match(image1, image2)
-      data = wrap_network_errors { facematch_service.facematch(facematch_body(image1, image2)) }
-      return unless data
-      scan_id_session[:facematch_pass] = facematch_pass?(data)
-    end
-
-    def facematch_body(image1, image2)
-      { 'Data':
-        { 'ImageOne': image1,
-          'ImageTwo': image2 },
-        'Settings':
-          { 'SubscriptionId': Figaro.env.acuant_assure_id_subscription_id } }.to_json
-    end
-
-    def fetch_face_from_document
-      assure_id.instance_id = scan_id_session[:instance_id]
-      data = wrap_network_errors { assure_id.face_image }
-      return if data.nil?
-      Base64.strict_encode64(data)
-    end
-
-    def process_selfie
-      liveness_body = request.body.read
-      liveness_data = wrap_network_errors { liveness_service.liveness(liveness_body) }
-      return unless liveness_data
-      unless selfie_live?(liveness_data)
-        render_json({})
-        return
-      end
-      JSON.parse(liveness_body)['Image']
-    end
-
-    def process_data_and_store_pii_in_session_if_document_passes(data, instance_id)
-      if data['Result'] == ACUANT_PASS
-        scan_id_session[:instance_id] = instance_id
-        scan_id_session[:pii] =
-          Idv::Utils::PiiFromDoc.new(data).call(current_user&.phone_configurations&.take&.phone)
-      end
-      data
-    end
-
-    def facematch_pass?(data)
-      data['IsMatch']
-    end
-
-    def selfie_live?(data)
-      data['LivenessResult']['LivenessAssessment'] == 'Live'
-    end
-
-    def assure_id
-      @assure_id ||= new_assure_id
-    end
-
-    def new_assure_id
-      (Rails.env.test? ? Idv::Acuant::FakeAssureId : Idv::Acuant::AssureId).new
-    end
-
-    def liveness_service
-      (Rails.env.test? ? Idv::Acuant::FakeAssureId : Idv::Acuant::Liveness).new
-    end
-
-    def facematch_service
-      (Rails.env.test? ? Idv::Acuant::FakeAssureId : Idv::Acuant::FacialMatch).new
-    end
 
     def return_good_document_if_throttled_else_increment
       render_json(GOOD_DOCUMENT) if Throttler::IsThrottledElseIncrement.call(*idv_throttle_params)
@@ -156,4 +71,3 @@ module Idv
     end
   end
 end
-# rubocop:enable Metrics/ClassLength
