+++ app/controllers/idv/scan_id_base_controller.rb
@@ -23,20 +23,6 @@ module Idv
       render json: data
     end
 
-    def wrap_network_errors
-      request_successful, data = yield
-      data = parse_if_json(data)
-      return data if request_successful
-      render json: {}, status: :service_unavailable
-      nil
-    end
-
-    def parse_if_json(data)
-      JSON.parse(data)
-    rescue JSON::ParserError
-      data
-    end
-
     def attempter_increment
       Throttler::Increment.call(*idv_throttle_params)
     end
